results/rebuttal_new_baselines_llmatic_missing/run-20231111-034044_LLMatic-zero_shot_donnemartin-system-design-oop-recipe_0_20-runs_log/recipe/2/LLMatic/recommendation.py
class Recommendation:
	def __init__(self, user, recipes):
		self.user = user
		self.recipes = recipes

	def generate_recommendations(self):
		# Mock implementation of recommendation generation based on user preferences and past activity
		# In a real-world scenario, this would involve complex algorithms and machine learning models
		recommended_recipes = [recipe for recipe in self.recipes if recipe.cuisine in self.user.preferences]
		return recommended_recipes

	def notify_user(self, new_recipes):
		# Mock implementation of user notification
		# In a real-world scenario, this would involve sending an email or push notification
		new_recipes_in_interest_areas = [recipe for recipe in new_recipes if recipe.cuisine in self.user.interest_areas]
		return new_recipes_in_interest_areas
